chore: some trivial reorganization and cleanup of code
