+++ src/obsidian_sync/scanner.py
